Fix prgressDialog and LogUtil
